refactor jsonrpc
