+++ app/services-manager.ts
@@ -47,55 +47,14 @@ import { Subscription } from 'rxjs/Subscription';
 import { Observable } from 'rxjs/Observable';
 import { GuestApiService } from 'services/guest-api';
 import { VideoEncodingOptimizationService } from 'services/video-encoding-optimizations';
+import {
+  IJsonRpcResponse,
+  IJsonRpcEvent,
+  IJsonRpcRequest,
+  E_JSON_RPC_ERROR
+} from 'services/jsonrpc';
 
 const { ipcRenderer } = electron;
-
-
-/**
- * @see http://www.jsonrpc.org/specification
- */
-export enum E_JSON_RPC_ERROR {
-  PARSE_ERROR = -32700,
-  INVALID_REQUEST = -32600,
-  METHOD_NOT_FOUND = -32601,
-  INVALID_PARAMS = -32602,
-  INTERNAL_JSON_RPC_ERROR = -32603,
-  INTERNAL_SERVER_ERROR = -32000
-}
-
-export interface IJsonRpcRequest {
-  jsonrpc: '2.0';
-  id: string | number;
-  method: string;
-  params: {
-    resource: string,
-    args?: any[],
-    fetchMutations?: boolean,
-    compactMode?: boolean
-  };
-}
-
-export interface IJsonRpcResponse<TResponse> {
-  jsonrpc: '2.0';
-  id?: string | number;
-  result?: TResponse;
-  error?: {
-    code: number;
-    message?: string
-  };
-  mutations?: IMutation[];
-}
-
-declare type TResourceType = 'HELPER' | 'SUBSCRIPTION' | 'EVENT';
-
-
-export interface IJsonRpcEvent {
-  _type: 'EVENT';
-  resourceId: string;
-  emitter: 'PROMISE' | 'STREAM';
-  data: any;
-  isRejected?: boolean;  // for PROMISE emitter only
-}
 
 
 export interface IMutation {
@@ -282,36 +241,6 @@ export class ServicesManager extends Service {
     } finally {
       return response;
     }
-  }
-
-
-  createErrorResponse(
-    options: { code: E_JSON_RPC_ERROR, id?: string|number, message?: string }
-  ): IJsonRpcResponse<any> {
-    return {
-      jsonrpc: '2.0',
-      id: options.id,
-      error: {
-        code: options.code,
-        message: E_JSON_RPC_ERROR[options.code] + (options.message ? (' ' + options.message) : ''),
-      }
-    };
-  }
-
-
-  createRequest(resource: any, method: string, ...args: any[]): IJsonRpcRequest {
-    const resourceId = resource.resourceId || resource.serviceName;
-    if (!resourceId) throw 'invalid resource';
-
-    return {
-      jsonrpc: '2.0',
-      id: ipcRenderer.sendSync('getUniqueId'),
-      method,
-      params: {
-        resource: resourceId,
-        args
-      }
-    };
   }
